requirements.txt
alembic==1.12.0
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.0.5
autopep8==2.0.1
backcall==0.2.0
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
colorama==0.4.4
cryptography==41.0.4
debugpy==1.6.0
decorator==5.1.1
distlib==0.3.6
distro==1.9.0
ecdsa==0.18.0
entrypoints==0.4
exceptiongroup==1.1.3
executing==0.8.3
fastapi==0.103.2
fastapi-auth0==0.5.0
filelock==3.9.0
greenlet==2.0.2
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.4
ipykernel==6.9.2
ipython==8.2.0
jedi==0.18.1
jupyter-client==7.1.2
jupyter-core==4.9.2
loguru==0.7.2
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
openai==1.12.0
parso==0.8.3
pickleshare==0.7.5
platformdirs==2.6.2
prompt-toolkit==3.0.28
psutil==5.9.0
psycopg2==2.9.9
pure-eval==0.2.2
pyasn1==0.5.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.11.2
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
pywin32==303
pyzmq==22.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
stack-data==0.2.0
starlette==0.27.0
tomli==2.0.1
tornado==6.1
tqdm==4.66.2
traitlets==5.1.1
typing_extensions==4.8.0
uvicorn==0.23.2
virtualenv==20.18.0
wcwidth==0.2.5
win32-setctime==1.1.0
